31/5/23

+++ swim/resources/views/scheduledwaste/pendingsw.blade.php
@@ -24,10 +24,8 @@
     body.loading .overlay{
         display: block;
     }
-
-
-        /* Add a CSS class to change the color to red */
-        .red-text {
+     /* Add a CSS class to change the color to red */
+     .red-text {
             color: red;
         }
    
@@ -66,14 +64,15 @@ $(document).ready(function() {
 <div class="card">
     <div class="card-header pb-0">
         <div class="row">
-            <div class=" {{  auth()->user()->category== 'Employee' ? 'col-lg-10 col-md-10 col-sm-10' : (request()->routeIs('swlist') ? 'col-lg-10 col-md-10 col-sm-10' : 'col-lg-12 col-md-12 col-sm-12') }}">
+            <div class=" {{  auth()->user()->category== 'Employee' ? 'col-lg-10 col-md-10 col-sm-10' : (request()->routeIs('pendingsw') ? 'col-lg-10 col-md-10 col-sm-10' : 'col-lg-12 col-md-12 col-sm-12') }}">
                 <nav class="">
                     <ul class="nav nav-tabs" id="myTab" role="tablist">
                         <li class="nav-item">
-                            <a class="nav-link {{ request()->routeIs('swlist') ? 'active' : '' }}" href="{{ route('swlist') }}" role="tab" aria-selected="true">List Of scheduledwaste</a>
+                        <a class="nav-link {{ request()->routeIs('swlist') ? '' : '' }}" href="{{ route('swlist') }}" role="tab" aria-selected="true">List Of scheduledwaste</a>
+
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link {{ request()->routeIs('pendingsw') ? '' : '' }}" href="{{ route('pendingsw') }}" role="tab" aria-selected="true">Pending Scheduled Waste</a>
+                            <a class="nav-link {{ request()->routeIs('pendingsw') ? 'active' : '' }}" href="{{ route('pendingsw') }}" role="tab" aria-selected="true">Pending Scheduled Waste</a>
                         </li>
                         <li class="nav-item">
                             <a class="nav-link {{ request()->routeIs('disposedsw') ? '' : '' }}" href="{{ route('disposedsw') }}" role="tab" aria-selected="true">Disposed Scheduled Waste</a>
@@ -84,7 +83,7 @@ $(document).ready(function() {
 
             @if( auth()->user()->category== "Employee")
 
-            @if(request()->routeIs('swlist'))
+            @if(request()->routeIs('pendingsw'))
             <div class="col-lg-2 col-md-2 col-sm-2" style="float: right;">
             <!-- wasteEmp nama apa2 sama dengan route kat web.php -->
                 <a class="btn btn-primary" style="float: right; width:100%;" role="button" href="{{ route('wasteEmp') }}">
@@ -120,27 +119,34 @@ $(document).ready(function() {
                         </tr>
                     </thead>
                     <tbody>
-                        
                     @foreach($wastelist as $index => $data)
-                    <tr id="row{{$data->id}}">
-                        <td>{{ $data->id }}</td>
-                        <td class="{{ $wasteData[$index]['diffInDays'] < 5 ? 'red-text' : '' }}">{{ $data->expiredDate }}</td>
-                        <td>{{ $data->wastecode }}</td>
-                        <td>{{ $data->statusDisposal }}</td>
-                        <td>
-                            @if(isset($wasteData[$index]))
-                                {{ $wasteData[$index]['diffInDays'] }} days
-                            @endif
-                        </td>
-                        <td>{{ $data->name }}</td>
-                        <td>
-                            <a type="button" class="btn btn-primary" href="{{ route('displaywaste', $data->id) }}">View</a>
-                            <button class="btn btn-danger" type="button" onclick="deleteItem(this)" data-id="{{ $data->id }}" data-name="{{ $data->wastecode }}">Delete</button>
-                        </td>                        
-                    </tr>
-                @endforeach
+                        @if($data->statusDisposal === 'Pending')
+                            <tr id="row{{$data->id}}">
+                                <td>{{ $data->id }}</td>
+                                <td class="{{ $wasteData[$index]['diffInDays'] < 5 ? 'red-text' : '' }}">{{ $data->expiredDate }}</td>
+                                <td>{{ $data->wastecode }}</td>
+                                <td>{{ $data->statusDisposal }}</td>
+                                <td>
+                                    @if(isset($wasteData[$index]))
+                                        {{ $wasteData[$index]['diffInDays'] }} days
+                                    @endif
+                                </td>
+                                <td>{{ $data->pic }}</td>
+                                <td>
+                                    <a type="button" class="btn btn-primary" href="{{ route('displaywaste', $data->id) }}">View</a>
+                                    <button class="btn btn-danger" type="button" onclick="deleteItem(this)" data-id="{{ $data->id }}" data-name="{{ $data->wastecode }}">Delete</button>
+                                </td>
+                            </tr>
+                        @endif
+                    @endforeach
 
+                    @if($wastelist->isEmpty())
+                        <tr>
+                            <td colspan="7" class="text-center">No records found.</td>
+                        </tr>
+                    @endif
                 </tbody>
+
                 
                 </table>
